FIX: missing enum serde tag FIX: `MessageSource` moved to `abi` module NEW: documentation section about threading
