+++ ton_client/client/src/abi/types.rs
@@ -1,7 +1,9 @@
-use crate::abi::Error;
+use crate::abi::{Error, ParamsOfEncodeMessage};
 use crate::error::ClientResult;
 use serde_json::Value;
 use ton_abi::{Token, TokenValue};
+use crate::{ClientContext, processing};
+use std::sync::Arc;
 
 #[derive(Serialize, Deserialize, Clone, Debug, ApiType, Default)]
 pub struct AbiHandle(u32);
@@ -86,3 +88,34 @@ impl FunctionHeader {
         Ok(Some(header))
     }
 }
+
+#[derive(Serialize, Deserialize, ApiType, Debug, Clone)]
+#[serde(tag="type")]
+pub enum MessageSource {
+    Encoded { message: String, abi: Option<Abi> },
+    EncodingParams(ParamsOfEncodeMessage),
+}
+
+impl MessageSource {
+    pub(crate) async fn encode(
+        &self,
+        context: &Arc<ClientContext>,
+    ) -> ClientResult<(String, Option<Abi>)> {
+        Ok(match self {
+            MessageSource::EncodingParams(params) => {
+                if params.signer.is_external() {
+                    return Err(processing::Error::external_signer_must_not_be_used());
+                }
+                let abi = params.abi.clone();
+                (
+                    crate::abi::encode_message(context.clone(), params.clone())
+                        .await?
+                        .message,
+                    Some(abi),
+                )
+            }
+            MessageSource::Encoded { abi, message } => (message.clone(), abi.clone()),
+        })
+    }
+}
+
